MB-11350 Rename ShipmentList/index.jsx to match component name
